Dispatching on Collection of missings

Trait IsEltypeSuperOfMissing defined using SimpleTraits.jl allows dispatching on any collection of eltype containing missings.

MissingStrategies.IsEltypeSuperOfMissing — Type

Trait which contains all types with Missing in its element type (eltype), excluding Any.

This allows dispatching on any Collection (Iterator, Vector, Tuple) that may contain missing elements.

See example of @handlemissings for an application.